Updating file names, class names and config option names for consistency and to prevent conflicts with other modules

+++ code/dev/UpdateSchemasTask.php
@@ -1,6 +1,6 @@
 <?php
 
-class UpdateDataSourcesTask extends BuildTask {
+class UpdateSchemasTask extends BuildTask {
 
     protected $title = 'Update Schema.org Data Sources';
 
@@ -40,23 +40,29 @@ class UpdateDataSourcesTask extends BuildTask {
             }
         }
 
-        // Process the schema.org Vocabulary Tree
-        echo "<h2>schema.org Vocabulary Tree</h2>";
-        $filename = "../" . Schema::getVocabularyFilePath();
-        $bytes = file_put_contents($filename, fopen("http://schema.org/docs/tree.jsonld", 'r'));
-        if($bytes === false) {
-            $error = true;
-            echo "<strong style=\"color: #FF0000\">Error: </strong>Unable to download schema.org Vocabulary Tree to " . $filename . "<br />";
-        } else {
-            echo "Downloaded schema.org Vocabulary Tree to " . $filename . "<br />";
+        if(!file_exists("../" . Schema::getVocabularyFilePath())) {
+            // Process the schema.org Vocabulary Tree
+            echo "<h2>schema.org Vocabulary Tree</h2>";
+            $filename = "../" . Schema::getVocabularyFilePath();
+            $bytes = file_put_contents($filename, fopen("http://schema.org/docs/tree.jsonld", 'r'));
+            if($bytes === false) {
+                $error = true;
+                echo "<strong style=\"color: #FF0000\">Error: </strong>Unable to download schema.org Vocabulary Tree to " . $filename . "<br />";
+            } else {
+                echo "Downloaded schema.org Vocabulary Tree to " . $filename . "<br />";
+            }
+            echo "<br />";
         }
-        echo "<br />";
         
+        // Update existing Schema's 
         foreach(Schema::get()->sort('SchemaOrgID', 'ASC') as $schema) {
             echo "<h2>schema.org " . $schema->Title . " Schema</h2>";
             $bytes = file_put_contents(
                 "../" . $schema->LocalDataSource,
-                fopen($schema->SchemaOrgURL . "." . Schema::$DataSource['format'], 'r')
+                fopen(
+                    $schema->SchemaOrgURL . "." . Schema::$DataSource['format'],
+                    'r'
+                )
             );
             if($bytes === false) {
                 $error = true;
@@ -76,6 +82,17 @@ class UpdateDataSourcesTask extends BuildTask {
                  */
             }
             echo "<br />";
+        }
+
+        // Create any missing schema's which should be enabled by default
+        $missing = array_diff(
+            Schema::getDefaultEnabledSchemas(),
+            Schema::get()->column('Title')
+        );
+        foreach($missing as $schema) {
+            $s = Schema::create();
+            $s->SchemaOrgID = 'schema:' . $schema;
+            $s->write();
         }
 
         echo (!$error) ? "<strong>Completed Successfully!</strong>" : "<strong>Completed With Errors</strong>";
